fix: build with clang>=15
In `ssl/statem/statem.c` the function fill_claim was used without a previous
declaration. This causes modern versions of clang to report an error and stop
the build process.

We add the missing declaration and include the rest of the claims interface,
where the C-types related to claims are defined.

see also https://github.com/tlspuffin/tlspuffin/pull/291 for tlspuffin upstream discussion